+++ src/xrGame/Inventory.cpp
@@ -62,7 +62,7 @@ CInventory::CInventory()
     m_fMaxWeight = pSettings->r_float("inventory", "max_weight");
     m_iMaxBelt = pSettings->read_if_exists<s32>("inventory", "max_belt", 5);
 
-    u16 sz;
+    size_t sz;
     const u16 tempSlotsCount = pSettings->read_if_exists<s16>("inventory", "slots_count", 10);
     if (tempSlotsCount > 0 && tempSlotsCount <= LAST_SLOT)
         sz = tempSlotsCount + 1;
