+++ src/xrGame/ai/monsters/ai_monster_squad_manager.cpp
@@ -30,9 +30,9 @@ void CMonsterSquadManager::register_member(u8 team_id, u8 squad_id, u8 group_id,
     // нет team - создать team, squad и group
     if (team_id >= team.size())
     {
-        team.resize(team_id + 1);
-        team[team_id].resize(squad_id + 1);
-        team[team_id][squad_id].resize(group_id + 1);
+        team.resize(static_cast<size_t>(team_id) + 1);
+        team[team_id].resize(static_cast<size_t>(squad_id) + 1);
+        team[team_id][squad_id].resize(static_cast<size_t>(group_id) + 1);
 
         for (u32 i = 0; i < group_id; i++)
             team[team_id][squad_id][i] = 0;
@@ -44,8 +44,8 @@ void CMonsterSquadManager::register_member(u8 team_id, u8 squad_id, u8 group_id,
     }
     else if (squad_id >= team[team_id].size())
     {
-        team[team_id].resize(squad_id + 1);
-        team[team_id][squad_id].resize(group_id + 1);
+        team[team_id].resize(static_cast<size_t>(squad_id) + 1);
+        team[team_id][squad_id].resize(static_cast<size_t>(group_id) + 1);
 
         for (u32 i = 0; i < group_id; i++)
             team[team_id][squad_id][i] = 0;
@@ -58,7 +58,7 @@ void CMonsterSquadManager::register_member(u8 team_id, u8 squad_id, u8 group_id,
     else if (group_id >= team[team_id][squad_id].size())
     {
         u32 prev_size = team[team_id][squad_id].size();
-        team[team_id][squad_id].resize(group_id + 1);
+        team[team_id][squad_id].resize(static_cast<size_t>(group_id) + 1);
 
         for (u32 i = prev_size; i < group_id; i++)
             team[team_id][squad_id][i] = 0;
